feat(codestyle): command - move embed/text generation out
