+++ src/commands/club/index.ts
@@ -1,7 +1,7 @@
-import { RichEmbed } from "discord.js";
+import { ICommand } from "../../interfaces/ICommand";
+import { consts } from "../../localization";
 
-import { ICommand } from "../interfaces/ICommand";
-import format, { consts } from "../localization";
+import { generateClubEmbed } from "./embed";
 
 module.exports = {
   name: "myclub",
@@ -33,22 +33,12 @@ module.exports = {
     }
 
     const homeclub_season = await homeclub.getSeason();
-    const seasons_count = homeclub.seasons_count ? `${format("season", homeclub.seasons_count)}` : "новый клуб";
-    const description = `Владелец - ${homeclub.owner_name} | ${format("participient", homeclub.members_count)} | ${seasons_count}`;
-    const points = `${homeclub_season.points}pt`;
-    const season_place = homeclub_season.rank ? `#${homeclub_season.rank}` : consts.noPlaceInTop;
 
-    const result = new RichEmbed()
-      .setColor("#0099ff")
-      .setTitle(`Клуб "${homeclub.name}"`)
-      .setDescription(description)
-      .addField("Общее количество очков", points)
-      .addField("Место в сезоне", season_place, true);
-
-    if (stage_data) {
-      result.addField(`Место в ${stage_data.number} этапе`, stage_data.place, true);
-    }
-
-    return message.channel.send(result);
+    const embed = generateClubEmbed({
+      homeclub,
+      homeclub_season,
+      stage_data
+    });
+    return message.channel.send(embed);
   }
 } as ICommand;
