+++ src/commands/farm/index.ts
@@ -1,26 +1,13 @@
 import { format as formatDate } from "date-fns";
 import { RichEmbed } from "discord.js";
 
-import { ICommand } from "../interfaces/ICommand";
-import { IStageSummoner } from "../interfaces/ISummoner";
-import format, { consts } from "../localization";
+import { ICommand } from "../../interfaces/ICommand";
+import { consts } from "../../localization";
 
-import { createPagedMessage } from "../helpers/discord";
-import { boldIF } from "../helpers/functions";
+import { createPagedMessage } from "../../helpers/discord";
 
-function formatDeficiencyMembers(embed: RichEmbed, summoners: IStageSummoner[], pointsNeeded: number, index_start = 1): RichEmbed {
-  const res = new RichEmbed(embed);
-  res.fields = [];
+import { formatDeficiencyMembers } from "./embed";
 
-  summoners
-    .forEach((member, i) => {
-      const title = boldIF(`${i + index_start}. ${member.summoner.summoner_name}`, i + index_start < 4);
-      const description = `${member.points}pt - ${format("game", member.games)} (нужно еще ${pointsNeeded - member.points}pt)`;
-      res.addField(title, description);
-    });
-
-  return res;
-}
 
 module.exports = {
   name: "myclubfarm",
@@ -76,8 +63,9 @@ module.exports = {
           const index_start = (page - 1) * count + 1;
 
           deficiency_slice = deficiency.slice((page - 1) * count, page * count);
-          result = formatDeficiencyMembers(template, deficiency_slice, index_start);
+          result = formatDeficiencyMembers(template, deficiency_slice, points, index_start);
           result.setFooter(`Страница ${page} • ${now}`);
+
           await deficiency_message.edit(result);
         },
         {
