+++ src/commands/season/embed.ts
@@ -1,0 +1,36 @@
+import { format as formatDate } from "date-fns";
+import { RichEmbed } from "discord.js";
+
+import LiveSeason from "../../models/LiveSeason";
+import Stage from "../../models/Stage";
+
+import { boldIF } from "../../helpers/functions";
+
+export function generateSeasonEmbed(
+  { live_season, stages }: { live_season: LiveSeason, stages: Stage[] }
+): RichEmbed {
+  const start_date = formatDate(live_season.start_date, "dd.MM.yyyy");
+  const end_date = formatDate(live_season.end_date, "dd.MM.yyyy");
+
+  const result = new RichEmbed()
+    .setColor("#0099ff")
+    .setTitle(`Информация о сезоне "${live_season.title}"`);
+
+  const season_description = live_season.isEnded() ?
+    `**Сезон окончен!** (Даты сезона: ${start_date} - ${end_date})`
+    : `Даты сезона: ${start_date} - ${end_date}`;
+
+  result.setDescription(season_description);
+
+  stages.forEach((stage) => {
+    const stage_start_date = formatDate(stage.start_date, "dd.MM.yyyy");
+    const stage_end_date = formatDate(stage.end_date, "dd.MM.yyyy");
+
+    const title = boldIF(`Этап ${stage.number}`, stage.is_live);
+    const description = boldIF(`${stage_start_date} - ${stage_end_date}`, stage.is_live);
+
+    result.addField(title, description);
+  });
+
+  return result;
+}
